Rename `featured-icons` to `featured-icon`
