+++ components/application/file-upload/file-upload-base.tsx
@@ -9,7 +9,7 @@ import { AnimatePresence, motion } from "motion/react";
 import { Button } from "@/components/base/buttons/button";
 import { ButtonUtility } from "@/components/base/buttons/button-utility";
 import { ProgressBar } from "@/components/base/progress-indicators/progress-indicators";
-import { FeaturedIcon } from "@/components/foundations/featured-icon/featured-icons";
+import { FeaturedIcon } from "@/components/foundations/featured-icon/featured-icon";
 import { cx } from "@/utils/cx";
 
 /**
@@ -23,6 +23,7 @@ export const getReadableFileSize = (bytes: number) => {
     const suffixes = ["B", "KB", "MB", "GB", "TB", "PB", "EB", "ZB", "YB"];
 
     const i = Math.floor(Math.log(bytes) / Math.log(1024));
+
     return Math.floor(bytes / Math.pow(1024, i)) + " " + suffixes[i];
 };
