+++ components/foundations/featured-icon/featured-icon.tsx
